Added package Workflow for workflow test Added package Util for small method

+++ ManningAutoTest/src/test/java/TestingData/HomeDeliveryTestData.java
@@ -55,7 +55,7 @@ public class HomeDeliveryTestData {
         this.isUpdateQty = value;
     }
 
-    public long getNewQty() {
+    public int getNewQty() {
         return newQty;
     }
